+++ Day24/Packaging.hpp
@@ -3,6 +3,6 @@
 
 #include <vector>
 
-unsigned int get_minimum_entaglement(const std::vector<unsigned int>& packages, unsigned int group_size);
+unsigned long get_minimum_entaglement(const std::vector<unsigned int>& packages, unsigned int group_size);
 
 #endif
